requirements.txt
streamlit
torch
faiss-cpu
numpy
PyPDF2
transformers
langchain
sentence-transformers
